Make JSONLoader work without passing onLoad method
The JSONLoader is built to support a manager passed in on creation. However, a normal case for using a manager is to keep track of how many files are being loaded against how many remain. Other loaders, such as TextureLoader, verify that the specific `onLoad` method was passed in before calling it; JSONLoader does not and throws when trying to call the function if it is not provided, even though the manager's is handling that.